+++ CHANGELOG.md
@@ -29,7 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ellipse2D: tangent_points
 
 #### edges.py
-- 
+- BSplineCurve: arc_to_nurbs and arc_to_nurbs_parameters
+
 #### wires.py
 - 
 #### surfaces.py
@@ -73,6 +74,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Unittests
 - Ellipse2D: test to the news functions
+- BSplineCurve: test to arc_to_nurbs and arc_to_nurbs_parameters
 
 ## v0.17.0 [Unreleased]
 ### New Features